requirements.txt
asgiref==3.6.0
Django==4.2
psycopg[binary]==3.1.8
whitenoise==6.2.0
djangorestframework==3.14.0
django-extensions==3.2.1
requests==2.28.2
pydantic==1.10.7
django-cors-headers==3.14.0